+++ exporter/prometheusexporter/prometheus_test.go
@@ -31,7 +31,6 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
-	"go.opentelemetry.io/collector/consumer/consumerdata"
 	"go.opentelemetry.io/collector/translator/internaldata"
 )
 
@@ -122,11 +121,11 @@ func TestPrometheusExporter_endToEnd(t *testing.T) {
 	require.NoError(t, exp.Start(context.Background(), &componenttest.NopHost{}))
 
 	// Should accumulate multiple metrics
-	md := internaldata.OCToMetrics(consumerdata.MetricsData{Metrics: metricBuilder(128, "metric_1_")})
+	md := internaldata.OCToMetrics(internaldata.MetricsData{Metrics: metricBuilder(128, "metric_1_")})
 	assert.NoError(t, exp.ConsumeMetrics(context.Background(), md))
 
 	for delta := 0; delta <= 20; delta += 10 {
-		md := internaldata.OCToMetrics(consumerdata.MetricsData{Metrics: metricBuilder(int64(delta), "metric_2_")})
+		md := internaldata.OCToMetrics(internaldata.MetricsData{Metrics: metricBuilder(int64(delta), "metric_2_")})
 		assert.NoError(t, exp.ConsumeMetrics(context.Background(), md))
 
 		res, err1 := http.Get("http://localhost:7777/metrics")
@@ -198,11 +197,11 @@ func TestPrometheusExporter_endToEndWithTimestamps(t *testing.T) {
 
 	// Should accumulate multiple metrics
 
-	md := internaldata.OCToMetrics(consumerdata.MetricsData{Metrics: metricBuilder(128, "metric_1_")})
+	md := internaldata.OCToMetrics(internaldata.MetricsData{Metrics: metricBuilder(128, "metric_1_")})
 	assert.NoError(t, exp.ConsumeMetrics(context.Background(), md))
 
 	for delta := 0; delta <= 20; delta += 10 {
-		md := internaldata.OCToMetrics(consumerdata.MetricsData{Metrics: metricBuilder(int64(delta), "metric_2_")})
+		md := internaldata.OCToMetrics(internaldata.MetricsData{Metrics: metricBuilder(int64(delta), "metric_2_")})
 		assert.NoError(t, exp.ConsumeMetrics(context.Background(), md))
 
 		res, err1 := http.Get("http://localhost:7777/metrics")